Moved SolverMeta and Solver to solvers/__init__.py and fixed appropriate imports in core.py
