+++ pywr/solvers/__init__.py
@@ -1,5 +1,38 @@
-#!/usr/bin/env python
-from ..core import Solver
+"""
+This module contains a Solver baseclass and several implemented subclasses.
+
+Solvers are used to with pywr.core.Model classes to solve the network
+allocation problem every time step.
+
+Currently there are only linear programme based solvers using,
+    - GLPK
+    - LPSolve55
+
+"""
+from six import with_metaclass
+
+
+class SolverMeta(type):
+    """Solver metaclass used to keep a registry of Solver classes"""
+    solvers = {}
+    def __new__(cls, clsname, bases, attrs):
+        newclass = super(SolverMeta, cls).__new__(cls, clsname, bases, attrs)
+        cls.solvers[newclass.name.lower()] = newclass
+        return newclass
+    @classmethod
+    def get_default(cls, ):
+        return cls.solvers['glpk']
+
+
+class Solver(with_metaclass(SolverMeta)):
+    """Solver base class from which all solvers should inherit"""
+    name = 'default'
+    def setup(self, model):
+        raise NotImplementedError('Solver should be subclassed to provide setup()')
+    def solve(self, model, timestep):
+        raise NotImplementedError('Solver should be subclassed to provide solve()')
+
+
 # Attempt to import solvers. These will only be successful if they are built correctly.
 try:
     from .cython_glpk import CythonGLPKSolver as cy_CythonGLPKSolver
